requirements.txt
pybullet>=3.2.5
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pyyaml>=6.0
robot-descriptions>=1.0.0
